pyproject.toml
[project]
name = "command-over-zmq"
version = "0.1.0"
description = "A clinet and server to run OS commands over socket using ZeroMQ"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyzmq>=26.4.0",
]
